backend/DappTrack_server/requirements.txt
fastapi[standard]
uvicorn==0.32.0
sqlalchemy==2.0.36
sqlalchemy_utils==0.41.2
asyncpg
pydantic_settings==2.6.1
cryptography==43.0.3
pyjwt==2.10.0
passlib==1.7.4
python-multipart==0.0.17
alembic
greenlet
boto3